refactor: refactor import statements for `TableRenderer` usage
- Import `TableRenderer` from a different location in `flytekit/deck/renderer.py`
- Remove `TableRenderer` from the imports in `flytekit-deck-standard/flytekitplugins/deck/__init__.py`
