+++ flytekit/deck/renderer.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, List, Optional
 
 from markdown_it import MarkdownIt
 from typing_extensions import Protocol, runtime_checkable
@@ -88,6 +88,53 @@ class SourceCodeRenderer:
         return f"<style>{css}</style>{html}"
 
 
+class TableRenderer:
+    """
+    Convert a pandas DataFrame into an HTML table.
+    """
+
+    def to_html(
+        self, df: pandas.DataFrame, header_labels: Optional[List] = None, table_width: Optional[int] = None
+    ) -> str:
+        # Check if custom labels are provided and have the correct length
+        if header_labels is not None and len(header_labels) == len(df.columns):
+            df = df.copy()
+            df.columns = header_labels
+
+        style = f"""
+            <style>
+                .table-class {{
+                    border: 1px solid #ccc;  /* Add a thin border around the table */
+                    border-collapse: collapse;
+                    font-family: Arial, sans-serif;
+                    color: #333;
+                    {f'width: {table_width}px;' if table_width is not None else ''}
+                }}
+
+                .table-class th, .table-class td {{
+                    border: 1px solid #ccc;  /* Add a thin border around each cell */
+                    padding: 8px;  /* Add some padding inside each cell */
+                }}
+
+                /* Set the background color for even rows */
+                .table-class tr:nth-child(even) {{
+                    background-color: #f2f2f2;
+                }}
+
+                /* Add a hover effect to the rows */
+                .table-class tr:hover {{
+                    background-color: #ddd;
+                }}
+
+                /* Center the column headers */
+                .table-class th {{
+                    text-align: center;
+                }}
+            </style>
+        """
+        return style + df.to_html(classes="table-class", index=False)
+
+
 class PythonDependencyRenderer:
     """
     PythonDependencyDeck is a deck that contains information about the library name and version.
@@ -100,8 +147,6 @@ class PythonDependencyRenderer:
         import json
         import subprocess
         import sys
-
-        from flytekitplugins.deck.renderer import TableRenderer
 
         from flytekit.loggers import logger
