+++ plugins/flytekit-deck-standard/flytekitplugins/deck/__init__.py
@@ -11,7 +11,6 @@ This package contains things that are useful when extending Flytekit.
    FrameProfilingRenderer
    MarkdownRenderer
    ImageRenderer
-   TableRenderer
 """
 
-from .renderer import BoxRenderer, FrameProfilingRenderer, ImageRenderer, MarkdownRenderer, TableRenderer
+from .renderer import BoxRenderer, FrameProfilingRenderer, ImageRenderer, MarkdownRenderer
