+++ plugins/flytekit-deck-standard/flytekitplugins/deck/renderer.py
@@ -1,5 +1,5 @@
 import warnings
-from typing import TYPE_CHECKING, List, Optional, Union
+from typing import TYPE_CHECKING, Optional, Union
 
 from flytekit import lazy_module
 from flytekit.types.file import FlyteFile
@@ -130,51 +130,6 @@ class ImageRenderer:
         img.save(buffered, format="PNG")
         img_base64 = base64.b64encode(buffered.getvalue()).decode()
         return f'<img src="data:image/png;base64,{img_base64}" alt="Rendered Image" />'
-
-
-class TableRenderer:
-    """
-    Convert a pandas DataFrame into an HTML table.
-    """
-
-    def to_html(self, df: pd.DataFrame, header_labels: Optional[List] = None, table_width: Optional[int] = None) -> str:
-        # Check if custom labels are provided and have the correct length
-        if header_labels is not None and len(header_labels) == len(df.columns):
-            df = df.copy()
-            df.columns = header_labels
-
-        style = f"""
-            <style>
-                .table-class {{
-                    border: 1px solid #ccc;  /* Add a thin border around the table */
-                    border-collapse: collapse;
-                    font-family: Arial, sans-serif;
-                    color: #333;
-                    {f'width: {table_width}px;' if table_width is not None else ''}
-                }}
-
-                .table-class th, .table-class td {{
-                    border: 1px solid #ccc;  /* Add a thin border around each cell */
-                    padding: 8px;  /* Add some padding inside each cell */
-                }}
-
-                /* Set the background color for even rows */
-                .table-class tr:nth-child(even) {{
-                    background-color: #f2f2f2;
-                }}
-
-                /* Add a hover effect to the rows */
-                .table-class tr:hover {{
-                    background-color: #ddd;
-                }}
-
-                /* Center the column headers */
-                .table-class th {{
-                    text-align: center;
-                }}
-            </style>
-        """
-        return style + df.to_html(classes="table-class", index=False)
 
 
 class GanttChartRenderer:
